pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "litdetect"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
license = { file = "LICENSE" }
keywords = ["detection", "YOLO", "PyTorch", "object-detection"]
authors = [
  { name = "wuxiheng", email = "abakir@qq.com" }
]
maintainers = [
  { name = "wuxiheng", email = "abakir@qq.com" }
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.12",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "absl-py==2.3.0",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.13",
    "aiosignal==1.3.2",
    "albucore==0.0.33",
    "albumentationsx>=2.0.9",
    "annotated-types==0.7.0",
    "antlr4-python3-runtime==4.9.3",
    "attrs==25.3.0",
    "certifi==2025.6.15",
    "charset-normalizer==3.4.2",
    "colorama>=0.4.6",
    "contourpy==1.3.2",
    "cycler==0.12.1",
    "faster-coco-eval==1.6.7",
    "filelock==3.18.0",
    "fonttools==4.58.4",
    "frozenlist==1.7.0",
    "fsspec==2025.5.1",
    "grpcio==1.73.0",
    "h5py~=3.14.0",
    "hf-xet==1.1.5",
    "huggingface-hub==0.33.0",
    "hydra-core==1.3.2",
    "idna==3.10",
    "jinja2==3.1.6",
    "joblib==1.5.1",
    "kiwisolver==1.4.8",
    "lightning==2.5.2",
    "lightning-utilities==0.14.3",
    "markdown==3.8.2",
    "markupsafe==3.0.2",
    "matplotlib==3.10.3",
    "mpmath==1.3.0",
    "multidict==6.5.1",
    "narwhals==1.44.0",
    "networkx==3.5",
    "notebook>=7.4.4",
    "numpy==2.3.1",
    "nvidia-cublas-cu12==12.6.4.1",
    "nvidia-cuda-cupti-cu12==12.6.80",
    "nvidia-cuda-nvrtc-cu12==12.6.77",
    "nvidia-cuda-runtime-cu12==12.6.77",
    "nvidia-cudnn-cu12==9.5.1.17",
    "nvidia-cufft-cu12==11.3.0.4",
    "nvidia-cufile-cu12==1.11.1.6",
    "nvidia-curand-cu12==10.3.7.77",
    "nvidia-cusolver-cu12==11.7.1.2",
    "nvidia-cusparse-cu12==12.5.4.2",
    "nvidia-cusparselt-cu12==0.6.3",
    "nvidia-nccl-cu12==2.26.2",
    "nvidia-nvjitlink-cu12==12.6.85",
    "nvidia-nvtx-cu12==12.6.77",
    "omegaconf==2.3.0",
    "onnxruntime>=1.22.0",
    "onnxsim>=0.4.36",
    "opencv-python-headless==4.11.0.86",
    "openpyxl>=3.1.5",
    "packaging==25.0",
    "pandas==2.3.0",
    "pillow==11.2.1",
    "plotly==6.1.2",
    "propcache==0.3.2",
    "protobuf==6.31.1",
    "pycocotools==2.0.10",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pyparsing==3.2.3",
    "python-dateutil==2.9.0.post0",
    "pytorch-lightning==2.5.2",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "requests==2.32.4",
    "safetensors==0.5.3",
    "scikit-image>=0.25.2",
    "scikit-learn==1.7.0",
    "scipy==1.16.0",
    "seaborn==0.13.2",
    "setuptools==80.9.0",
    "simpleitk>=2.5.2",
    "simsimd==6.4.9",
    "six==1.17.0",
    "stringzilla==3.12.5",
    "sympy==1.14.0",
    "tabulate==0.9.0",
    "tensorboard==2.19.0",
    "tensorboard-data-server==0.7.2",
    "tensorboardx==2.6.4",
    "termcolor==3.1.0",
    "threadpoolctl==3.6.0",
    "timm==1.0.15",
    "torch==2.7.1",
    "torchio>=0.20.15",
    "torchmetrics==1.7.3",
    "torchvision==0.22.1",
    "tqdm==4.67.1",
    "triton==3.3.1",
    "typing-extensions==4.14.0",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "wcwidth>=0.2.13",
    "werkzeug==3.1.3",
    "yarl==1.20.1",
]

[tool.setuptools]
package-dir = {"" = "src"}

[project.urls]
homepage = "https://github.com/AbaK1r/LitDetect"
repository = "https://github.com/AbaK1r/LitDetect"
documentation = "https://github.com/AbaK1r/LitDetect#readme"
bug-tracker = "https://github.com/AbaK1r/LitDetect/issues"

[project.scripts]
litdetect-train = "scripts.train:main"
